Add syntactic analysis script
The function of the file "syntax_analyse_clean_dataCL.py" is to augment the dataframe with syntactic (constituency) analysis. It is a python file executable via the command line: It takes the clean_text column from the clean data as an as input, adds the columns 'tree' and 'root', and outputs the resulting dataframe to the desired location. While the 'root' column is the only one used in this study, the 'tree' column stores the entire constituent information in bracket notation, which may be desirable for follow-up more in-depth analyses.