Add feature extraction script
The function of the file "feature_extract_clean_dataCL.py" is to augment the dataframe with a list of linguistic features taken from or inspired by the literature. It is a python file executable via the command line. It adds many feature columns and outputs the resulting dataframe to the desired location